+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 * Deployment: Send a non 0 exit code when a deployment fails [#563](https://github.com/Scalingo/cli/pull/563)
 * Bugfix: fix support for `addon_updated` and `start_region_migration` event type [#558](https://github.com/Scalingo/cli/pull/558)
 * Bugfix: fix author of `restart` and `edit_variable` when it's an addon [#558](https://github.com/Scalingo/cli/pull/558)
+* Add log drains add command
+  [#561](https://github.com/Scalingo/cli/pull/561)
+  [go-scalingo #164](https://github.com/Scalingo/go-scalingo/pull/164)
+
 
 ### 1.16.8